requirements.txt
Flask==3.0.3
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.24.0
gunicorn==22.0.0